requirements.txt
celery==5.3.0
django==4.2.7
djangorestframework==3.14.0
mysqlclient==2.1.1
redis==5.0.1
pytest==7.4.3
